chore: 🤖 move source code to /src and add absolute imports
