--- packages/app/graphql/mutations/index.ts
@@ -1,7 +1,0 @@
-import { gql } from '@apollo/client'
-
-export const INIT_DRONE = gql`
-  mutation initDrone($start: DestinationInput!, $stop: DestinationInput!) {
-    initDrone(start: $start, stop: $stop)
-  }
-`
