+++ packages/app/src/App.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { ApolloProvider } from '@apollo/client'
 import { ThemeProvider } from 'styled-components/native'
-import client from '~/graphql/apolloClient'
 import Home from '~/views/Home'
 import theme from '~/styles/theme'
+import client from '~/graphql/apolloClient'
 
 const App = () => {
   return (
